==> Step 1     Scrapping The Packages.ipynb <==
import requests
from bs4 import BeautifulSoup
import numpy as np
import pandas as pd
from tqdm.notebook import tqdm 

tour_type = []
tour_name = []
destinations_link = []
destinations = []
age = []
tour_type_des = []
vendor = []
Duration = []
price = []
for i in tqdm(range(31)):
    print('Started')
    url =  "https://www.tourradar.com/d/india?page={}".format(i)
    content = requests.get(url)
    soup = BeautifulSoup(content.text, 'html.parser')
    tour_data  = soup.find_all('li', class_ = 'tour')
    for i in tour_data:
        try:
            tour_type.append(i.find('a' , class_ = "ao-serp-tour__travel-style-link").text.strip())
        except:
             tour_type.append(np.nan)
        tour_name.append(i.find('h4' , class_ = "js-ao-serp-tour__name").text.strip())   
        x = i.find('h4' , class_ = "js-ao-serp-tour__name")
        y = str(x).split('data-id=')[1]
        z = y.split('"')[1]
        destinations_link.append(int(z))

        ## Paragarph text with same tag so deal with it with some specific method
        text= i.find('dl').text

        ## Getting Destinations
        idx1 = text.find("Destinations")
        idx2 = text.find("Age Range")
        destinations.append(text[idx1 + len("Destinations") + 1: idx2])

        ##Getting Age
        idx3 = text.find("Age Range")
        idx4 = text.find("year")
        age.append( text[idx3 + len("Age Range") + 1: idx4])

        ##Getting tour type
        idx5 = text.find("Style")
        idx6 = text.find("Operated")
        tour_type_des.append(text[idx5 + len("Style") : idx6])

        ## Getting Vendor
        idx5 = text.find("Operator")
        vendor.append(text[idx5 + len("Operator") : ])

        ## Getting Durations
        try:
            Duration.append(i.find('dd', class_ = 'br__price-wrapper-info-description').text)
        except:
            Duration.append(np.nan)

        ## Getting the price    
        try:
            price.append(i.find('span', class_ = 'js-br__price-wrapper-price-description-value br__price-wrapper-price-description-value').text)

        except:
            price.append(np.nan) 

        ## Creating a dictionary
    d = {'Category': tour_type, 'name' : tour_name, 'destinations' : destinations, 'age_group' : age,
             'tour_type_des' :tour_type_des, 'Vendors': vendor, 'Duration' : Duration, 'price':price,
         'Destination_link' : destinations_link 
            }
    df = pd.DataFrame(d)
    

print(df.shape)
india_csv_data = df.to_csv('india_Package.csv', index = True)

df.head()
