==> Step 2    Review Scrapping.ipynb <==
import pandas as pd
import requests
from bs4 import BeautifulSoup

df = pd.read_csv('india_Package.csv')
df.head()

df.shape

len(df['Destination_link'].unique())

bolean = df['Destination_link'].duplicated()

bolean

df = df[~bolean]

df.shape

from tqdm.notebook import tqdm 

comment = []
destination = []
for j in tqdm(range(0, len(df))):
    url =  "https://www.tourradar.com/t/{}#p=1_".format(df['Destination_link'][j])
    content = requests.get(url)
    soup = BeautifulSoup(content.text, 'html.parser')
    tour_review  = soup.find_all('div', class_ = 'js-am-tour-reviews__review-item-body')
    for i in tour_review:
        x = i.text.strip()
        comment.append(x)
        destination.append(df['Destination_link'][j])
    

d = {'Destination_link' : destination , 'Review': comment}
df1 = pd.DataFrame(d)
df1.head(50)

India_Review = df1.to_csv('IndiaReview.csv', index = True)

for i in df['Destination_link']:
    url =  "https://www.tourradar.com/t/{}#p=1_".format(i)
    content = requests.get(url)
    soup = BeautifulSoup(content.text, 'html.parser')
    tour_review  = soup.find_all('div', class_ = 'js-am-tour-reviews__review-item-body')
    comment = []
    for i in tour_review:
        x = i.text.strip()
        comment.append(x)
        
    

for i in df['Destination_link']:
    print(i)

url = "https://www.tourradar.com/t/115430#p=1_"
content = requests.get(url)
soup = BeautifulSoup(content.text, 'html.parser')
tour_review  = soup.find_all('div', class_ = 'js-am-tour-reviews__review-item-body')
comment = []
for i in tour_review:
    ## print(i.text.strip())
    x = i.text.strip()
    comment.append(x)

len(comment)

content = requests.get(url)

soup = BeautifulSoup(content.text, 'html.parser')

tour_review  = soup.find_all('div', class_ = 'js-am-tour-reviews__review-item-body')

tour_review

len(tour_review)

for i in tour_review:
    print(i.text.strip())

tour_type = []
tour_name = []
comments = []
link = []
for j in range(1,2):
    url =  "https://www.tourradar.com/d/india?pages={}".format(j)
    content = requests.get(url)
    soup = BeautifulSoup(content.text, 'html.parser')
    tour_data  = soup.find_all('li', class_ = 'tour')
    for i in tour_data:
        try:
            tour_type.append(i.find('a' , class_ = "ao-serp-tour__travel-style-link").text.strip())
        except:
             tour_type.append(np.nan)
        tour_name.append(i.find('h4' , class_ = "js-ao-serp-tour__name").text.strip())
        text = i.find('a', class_= "blank tourLink")
        print(text)
        sub = 'href'
        ## Getting Link for Comments
        index1 = text.find(sub)
        

       

        ## Creating a dictionary
    d = {'Category': tour_type, 'name' : tour_name}
    df = pd.DataFrame(d)
    

<a href="/t/5505#p=1_" class="blank tourLink" data-id="5505"><h4 itemprop="name" class="js-ao-serp-tour__name" data-id="5505">Golden Triangle</h4></a>

tour_type = []
tour_name = []
comments = []
link = []
j = 0
url =  "https://www.tourradar.com/d/india?pages=1"
content = requests.get(url)
soup = BeautifulSoup(content.text, 'html.parser')
tour_data  = soup.find_all('li', class_ = 'tour')
for i in tour_data:
    try:
        tour_type.append(i.find('a' , class_ = "ao-serp-tour__travel-style-link").text.strip())
    except:
        tour_type.append(np.nan)
    tour_name.append(i.find('h4' , class_ = "js-ao-serp-tour__name").text.strip())
    link.append(i.find('a', class_= "blank tourLink"))
text = link[0]
print(text)
sub = 'href'
## Getting Link for Comments


type(text)

text = str(text)

idx5 = text.find("href=")
idx6 = text.find("\">")
print(text[idx5 + len("href=\"") :idx6 ])
tour_id = text[idx5 + len("href=\"") :idx6 ]

url = "https://www.tourradar.com{}#p=1_".format(tour_id)
content = requests.get(url)

content

url = "https://www.tourradar.com{}#p=1_".format(tour_id)
content = requests.get(url)
soup = BeautifulSoup(content.text, 'html.parser')
tour_review  = soup.find_all('div', class_ = 'js-am-tour-reviews__review-item-body')
for i in tour_review:
    print(i.text.strip())

dir(text)
