==> Step 4    Review Data Set combine .ipynb <==
import pandas as pd 

df1 = pd.read_csv('India Sentimental Scores for Reviews.csv')[['Destination_link','Review','Sentiment Score']]
df2 = pd.read_csv('USA Sentimental Scores for Reviews.csv')[['Destination_link','Review','Sentiment Score']]
df3 = pd.read_csv('NorwayReviewScore.csv')
df3 = df3.rename(columns={"Scores": "Sentiment Score"})
df4 = pd.read_csv('Srilamka Sentimental Scores for Reviews.csv')[['Destination_link','Review','Sentiment Score']]
df5 = pd.read_csv('Canada Sentimental Scores for Reviews.csv')[['Destination_link','Review','Sentiment Score']]
df6 = pd.read_csv('Egypt Sentimental Scores for Reviews.csv')[['Destination_link','Review','Sentiment Score']]
df7 = pd.read_csv('Pakistan Sentimental Scores for Reviews.csv')[['Destination_link','Review','Sentiment Score']]

print(df1.shape)
print(df2.shape)
print(df3.shape)
print(df4.shape)
print(df5.shape)
print(df6.shape)
print(df7.shape)

df1.head()

df2.head()

df3.head()

df4.head()

df5.head()

df6.head()

df7.head()

### Categories in Each Data Set

df1.groupby('Sentiment Score').count()

df2.groupby('Sentiment Score').count()

df3.groupby('Sentiment Score').count()

df4.groupby('Sentiment Score').count()

df5.groupby('Sentiment Score').count()

df6.groupby('Sentiment Score').count()

df7.groupby('Sentiment Score').count()

# Conbine all the data set 

df = pd.concat([df1,df2,df3,df4,df5,df6,df7], axis = 0)
df.head()

## Shape of the combined Data set 

df.shape

### Count of each Category 

df.groupby('Sentiment Score').count()

df.to_csv('Combine Review Data set.csv')
