==> Step 5     Analysis on the Review Category for each label.ipynb <==
import pandas as pd 

df = pd.read_csv('India Sentimental Scores for Reviews.csv')[['Destination_link','Review','Sentiment Score']]

df.head()

df.groupby('Sentiment Score').count()['Review']

df1 = pd.read_csv('USA Sentimental Scores for Reviews.csv')[['Destination_link', 'Review', 'Sentiment Score']]
df1.head()

df1 = df1[df1['Sentiment Score'] <= 3]

final_df = pd.concat([df,df1], axis = 0)

final_df.shape

final_df.groupby('Sentiment Score').count()['Review']

df2 = pd.read_csv('NorwayReviewScore.csv')
df2.head()

df2 = df2.rename(columns={"Scores": "Sentiment Score"})
df2.head()

df2 = df2[df2['Sentiment Score'] <= 4]
df2

final_df = pd.concat([final_df,df2], axis = 0)

final_df.groupby('Sentiment Score').count()['Review']

df3 = pd.read_csv('Srilamka Sentimental Scores for Reviews.csv')
df3.head()

df3 = df3[df3['Sentiment Score'] <= 4]
df3

df3.groupby('Sentiment Score').count()['Review']

final_df = pd.concat([final_df,df3], axis = 0)
final_df.groupby('Sentiment Score').count()['Review']

df1 = pd.read_csv('Canada Sentimental Scores for Reviews.csv')[['Destination_link', 'Review', 'Sentiment Score']]
df1.head()

df1 = df1[df1['Sentiment Score'] <= 4]
df1

df1[df1['Sentiment Score'] <= 1]

len(df1[df1['Sentiment Score'] == 2])

final_df = pd.concat([final_df,df1], axis = 0)
final_df.groupby('Sentiment Score').count()['Review']

final_df.shape

### Let's make the final df with each category have 200 records 

df1 = final_df[final_df['Sentiment Score'] == 1][['Destination_link', 'Review', 'Sentiment Score']]
print(df1.shape)
df1.head()

df2 = final_df[final_df['Sentiment Score'] == 2][['Destination_link', 'Review', 'Sentiment Score']]
print(df2.shape)
df2.head()

df3 = final_df[final_df['Sentiment Score'] == 3][['Destination_link', 'Review', 'Sentiment Score']]
print(df3.shape)
df3.head()

df3 = df3.iloc[:250]
df3.shape

df4 = final_df[final_df['Sentiment Score'] == 4][['Destination_link', 'Review', 'Sentiment Score']]
print(df4.shape)
df4.head()

df4 = df4.iloc[:250]
df4.shape

df5 = final_df[final_df['Sentiment Score'] == 5][['Destination_link', 'Review', 'Sentiment Score']]
print(df5.shape)
df5.head()

df5 = df5.iloc[:250]
df5.shape

### Combine them all into One Data Frame 

df = pd.concat([df1,df2,df3,df4,df5], axis = 0)
df.head()

df.shape

df.to_csv('Review Balanced Data set.csv')
