==> Step 7    Sentimental Analysis Using Neural Network.ipynb <==
import pandas as pd
df = pd.read_csv("Combine Review Data set.csv")[['Destination_link','Review','Sentiment Score']]
df.head(4)

df.shape

df.isnull().sum()

x = df['Review']
y = df['Sentiment Score']

print(x.shape)
print(y.shape)

## Now we will apply some preprocessing on the x 

x

## importing libraries for text preprocessing

import nltk
import re
from nltk.corpus import stopwords

nltk.download('stopwords')

from nltk.stem.porter import PorterStemmer

## Creating obj of the porter stemmer
stemmer = PorterStemmer()

# Applying preprocessing on data

- Removing unwanted Characters
- eliminate stop words
- lower case
- Stemming

## for showing the progress bcz it's takes time 
from tqdm import tqdm

corpus = []
for i in tqdm(range(0, len(x))):
    review = re.sub('[^a-zA-Z]', " ", x[i])
    review = review.lower()
    review = review.split()
    
    review = [ stemmer.stem(words) for words in review if not words in stopwords.words('english')]
    review = ' '.join(review)
    corpus.append(review)

corpus[5]

!pip install tensorflow-gpu

import tensorflow as tf

## Creating Embedding layer

from tensorflow.keras.preprocessing.sequence import pad_sequences
from tensorflow.keras.layers import Embedding
from tensorflow.keras.models import Sequential
from tensorflow.keras.preprocessing.text import one_hot
from tensorflow.keras.layers import LSTM
from tensorflow.keras.layers import Dense

vocab_size = 1000000

## Getting indexes after applying the one hot ecoding using keras

one_hot_rep = [ one_hot(word, vocab_size) for word in corpus]

## Finding the highest len of the review in the corpus

len(one_hot_rep[34])

len(one_hot_rep[3400])

len(one_hot_rep[5000])

max_list = max(one_hot_rep, key=len)

# Print the result
print(len(max_list))

lenght = 46
for i in tqdm(range(1, len(corpus))):
    if lenght < len(one_hot_rep[i]):
        lenght = len(one_hot_rep[i])
        print(lenght, " and the index is:" , i)

corpus[0]

lenght

### let's find the mean here for the sentence len

sum_len = 0
for i in tqdm(range(0, len(corpus))):
    sum_len = sum_len + len(one_hot_rep[i])

sum_len / len(corpus)

## So here we considered the max len 130 and the the review above that len will get discarded

length = 130

### Applying pedding to keep the size of the input same that's why we above find the len of the sentence

padding_data = pad_sequences(one_hot_rep, padding = 'post', maxlen = length)

padding_data[35]

### Embedding Layer
It performs embedding operations in input layer. It is used to convert positive into dense vectors of fixed size. Its main application is in text analysis. 

### As we get tense vocab, now we need to convert it into the vectors.

### for this, we will be taking each word and than representing it with the number of features (let say we pass the len 8, than each word is represented by 8 different values which are actually the features)

embedding_layer_features = 170

## Using Bidirectional LSTM here

from tensorflow.keras.layers import Bidirectional


model = Sequential()
model.add(Embedding( input_dim = vocab_size,output_dim = embedding_layer_features, input_length = length))
model.add(Bidirectional(LSTM(200)))
model.add(Dense(1, activation='sigmoid'))
model.compile(loss='binary_crossentropy', optimizer = "adam", metrics = ['accuracy'])
print(model.summary())

len(padding_data), y.shape

x = padding_data

## Splitting of the data

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(x, y, test_size=0.33, random_state=42)

## Training of the model

model.fit(X_train,y_train,validation_data=(X_test,y_test),epochs=10,batch_size=64)

y_pred=model.predict(X_test)

import numpy as np
b=np.where(y_pred > 0.6, 1,0) ##AUC ROC Curve 

from sklearn.metrics import confusion_matrix
confusion_matrix(y_test,y_pred)

from sklearn.metrics import classification_report
print(classification_report(y_test,y_pred))

from sklearn.metrics import accuracy_score
accuracy_score(y_test,y_pred)

y_test

y_pred[1]
