==> Step 8    Destination Recommendation.ipynb <==
import pandas as pd 

df = pd.read_csv('Pakistan Tourist Destinations.csv')
df.head()

df.head(30)

df.shape

df['Desc'][3]

df.isnull().sum()

combined_features = df['_key'] + " "+ df['Desc'] + " "+ df['category'] +" "+ df['district']

combined_features

# Applying TFIDF Technique

import difflib
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity

vectorizer = TfidfVectorizer()
feature_vector = vectorizer.fit_transform(combined_features)
print(feature_vector)

similarity = cosine_similarity(feature_vector)
print(similarity)

## Get the name of place from the user 

desti_name = input("Enter the name of Place : ")

destinations = df['_key'].tolist()
len(destinations)

find_closed_match = difflib.get_close_matches(desti_name, destinations)
print(find_closed_match)

## Finding the closet match 

closed_match = find_closed_match[0]
closed_match

## Getting index 

index_of_desti = df[df['_key'] == closed_match].index.values[0]
print(index_of_desti)

## Getting Similar items 

similarity_score = list(enumerate(similarity[index_of_desti]))

# means that it is generated for each and every destination in the dataset
len(similarity_score)

# Sorting the movies on the based of similarity Score

sorted_desti_list = sorted(similarity_score, key = lambda x : x[1], reverse = True)
sorted_desti_list

print("Destination Suggested to you are : ")
i = 1
name = []
for desti in sorted_desti_list:
    index = desti[0]
    name.append(df.iloc[index]['_key'])

print("Top 5 similar Items:")
for i in range(5):
    print(i+1, ": ", name[i])

 df[df['_key'] == 'Lahore Fort']['Desc']

## Function for the generation of the Recommendations

def Recommendation(desti_name):
    find_closed_match = difflib.get_close_matches(desti_name, destinations)
    closed_match = find_closed_match[0]
    index_of_desti = df[df['_key'] == closed_match].index.values[0]
    similarity_score = list(enumerate(similarity[index_of_desti]))
    sorted_desti_list = sorted(similarity_score, key = lambda x : x[1], reverse = True)
    print("Destination Suggested to you are : ")
    i = 1
    name = []
    for desti in sorted_desti_list:
        index = desti[0]
        name.append(df.iloc[index]['_key'])

    print("Top 5 similar Items:")
    names = []
    for i in range(1,6):
        # print(i, ": ", name[i])  
        names.append(name[i])
    
    return names 

desti_name = input("Enter the name of Place : ")
Recommendation(desti_name)

import pickle 

## Pickling the modules 

pickle.dump(destinations,open('destinations.pkl', 'wb'))
pickle.dump(df,open('df.pkl', 'wb'))

pickle.dump(similarity,open('similarity.pkl', 'wb'))

similarity
